Server-side configuration code made plain
