+++ api/swagger/swagger.yaml
@@ -4,41 +4,33 @@ info:
   title: Hello World App
 # during dev, should point to your local machine
 host: localhost:3000
-# basePath prefixes all resource paths 
 basePath: /
 # 
 schemes:
   # tip: remove http to make production-grade
   - http
   - https
-# format of bodies a client can send (Content-Type)
 consumes:
   - application/json
-# format of the responses to the client (Accepts)
 produces:
   - application/json
 paths:
   /hello:
-    # binds a127 app logic to a route
     x-swagger-router-controller: hello-world
     post:
       description: Returns 'Hello' to the caller
-      # used as the method name of the controller
       operationId: sayHello
       responses:
         "200":
           description: Success
           schema:
-            # a pointer to a definition
             $ref: "#/definitions/HelloWorldResponse"
-        # responses may fall through to errors
         default:
           description: Error
           schema:
             $ref: "#/definitions/ErrorResponse"
   /swagger:
     x-swagger-pipe: swagger_raw
-# complex objects have schema definitions
 definitions:
   HelloWorldResponse:
     required:
