--- app/www
@@ -1,28 +1,0 @@
-// dependencies
-
-var q = require('q');
-var mongoose = require('mongoose');
-
-var serverBuilder = require('./core/server.factory.js');
-var portTool = require('./tools/port.tool.js');
-
-// exports
-
-module.exports = {
-    start: _start
-};
-
-// initialization
-
-// private methods
-
-function _start(port, app) {
-    var normalizedPort = portTool.normalizePort(port);
-
-    // Plugging mongoose in Promises Library (http://mongoosejs.com/docs/promises.html)
-    mongoose.Promise = q.Promise;
-
-    app.set('port', normalizedPort);
-
-    return serverBuilder.startServer(normalizedPort, app);
-}
